Reorganize the navigation of screens properly (#13)
* Define a sealed calss for navigation routes

* Move `SplashScreen()` to a separate file

* Update the navigations between screens
