--- app/src/main/java/com/it235/nureserved/AuthScreen.kt
@@ -1,26 +1,0 @@
-package com.it235.nureserved
-
-import androidx.compose.runtime.Composable
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import com.it235.nureserved.ui.authscreenui.LoginScreen
-import com.it235.nureserved.ui.authscreenui.SignUpScreen
-
-// Navigation for authentication screens
-@Composable
-fun AuthScreenNav() {
-    val navController = rememberNavController()
-
-    NavHost(
-        navController = navController,
-        startDestination = Routes.LoginScreen,
-    ) {
-        composable(route = Routes.LoginScreen) {
-            LoginScreen(navController)
-        }
-        composable(route = Routes.SignUpScreen) {
-            SignUpScreen(navController)
-        }
-    }
-}
